Move rename AvroSinkValueParameter file to AvroSchemaBasedParameter, move it out of sink package
